+++ pytcs/pytcs.py
@@ -2,6 +2,7 @@
 from __future__ import annotations
 
 import gzip
+import importlib
 from collections.abc import ItemsView, Iterator, ValuesView
 from dataclasses import dataclass, field
 from datetime import datetime
@@ -227,6 +228,8 @@ class ScopeFile:
 
         if backend == "pandas":
             data_dict = self._read_pandas(usecols, native_dtypes)
+        elif backend == "pyarrow":
+            data_dict = self._read_pyarrow(usecols, native_dtypes)
         elif backend == "datatable":
             if native_dtypes:
                 warn(
@@ -593,7 +596,68 @@ class ScopeFile:
             compression=self._compression,
         )
 
-        # self._df = df  # debug
+        if native_dtypes:
+            tc3_dtypes = get_tc3_dtypes()
+            dtypes_np = {
+                v.value_col: tc3_dtypes[v.info.get("Data-Type", "REAL64")][0]
+                for v in self._channels.values()
+            }
+            dtypes_times = {k: np.float64 for k in self._get_time_cols()}
+            dtypes_np.update(dtypes_times)
+            data_dict = {k: df[k].dropna().to_numpy(dtypes_np[k]) for k in df}
+        else:
+            data_dict = {k: df[k].dropna().to_numpy(np.float64) for k in df}
+
+        return data_dict
+
+    def _read_pyarrow(
+        self, usecols: list[int], native_dtypes: bool
+    ) -> dict[str, np.ndarray]:
+        if not importlib.util.find_spec("pyarrow"):
+            warn(
+                "'pyarrow' backend not found, using default pandas implementation.",
+                UserWarning,
+                stacklevel=2,
+            )
+            return self._read_pandas(usecols, native_dtypes)
+
+        """Read data into dictionary using the pandas backend."""
+        if isinstance(self._file, IOBase):  # read open streams from beginning
+            self._file.seek(0)
+
+        if self._delimiter not in (" ", ","):
+            warn(
+                f"""Unsupported delimiter '{self._delimiter}' for pyarrow backend. """
+                """Falling pack to pandas implementation.""",
+                UserWarning,
+                stacklevel=2,
+            )
+            return self._read_pandas(usecols, native_dtypes)
+        if self.decimal not in ("."):
+            warn(
+                f"""Unsupported decimal '{self._decimal}' for pyarrow backend. """
+                """Falling pack to pandas implementation.""",
+                UserWarning,
+                stacklevel=2,
+            )
+            return self._read_pandas(usecols, native_dtypes)
+
+        df = pd.read_csv(
+            self._file,
+            delimiter=self._delimiter,
+            skiprows=self._header_lines,
+            # decimal=self._decimal,
+            header=None,
+            usecols=usecols,
+            names=self._get_cols(),
+            index_col=False,
+            encoding=self._encoding,
+            na_values=[" ", "EOF"],
+            skip_blank_lines=True,
+            engine="pyarrow",
+            # low_memory=False, # unsupported with pyarrow
+            compression=self._compression,
+        )
 
         if native_dtypes:
             tc3_dtypes = get_tc3_dtypes()